+++ pwa-components-viewme/src/global/constants.ts
@@ -1,1 +1,1 @@
-export const APPVERSION:string ="0.0.3-2";
+export const APPVERSION:string ="0.0.3-4";